news_sentiment/requirements.txt
flask
pandas
textblob
mysql-connector-python
pymysql
PyMySQL[rsa]
cryptography
requests
SQLAlchemy
langdetect